feature(back): drop opentdb requests, add getting questions from db

+++ back/src/main/java/com/gehtsoft/dto/quiz/CheckAnswersRequestBody.java
@@ -12,7 +12,7 @@ public class CheckAnswersRequestBody {
 
     @Data
     public static class Answer {
-        private Long questionId;
+        private int questionId;
         private int answerIdx;
     }
 }
